# ProductAPI/src/main/java/org/ecommerce/productapi/dto/order/OrderItemResponse.java
package org.ecommerce.productapi.dto.order;

import org.ecommerce.productapi.dto.perfume.PerfumeResponse;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class OrderItemResponse {
    private Long id;
    private Long amount;
    private Long quantity;
    private PerfumeResponse perfume;
}